removed warnings

+++ include/wrench/services/compute/cloud/CloudService.h
@@ -119,7 +119,7 @@ namespace wrench {
 
 
         /** @brief A map of VMs described by the VM actor, the actual compute service, and the total number of cores */
-        std::map<std::string, std::tuple<simgrid::s4u::VirtualMachine *, std::unique_ptr<ComputeService>, int>> vm_list;
+        std::map<std::string, std::tuple<simgrid::s4u::VirtualMachine *, std::unique_ptr<ComputeService>, unsigned long>> vm_list;
     };
 
 }
